# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dominion Realm API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/chapters/{chapter_id}/telemetry": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Chapter Telemetry",
        "operationId": "chapter_telemetry_chapters__chapter_id__telemetry_get",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Chapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterTelemetryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}/telemetry": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Book Telemetry",
        "operationId": "book_telemetry_books__book_id__telemetry_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookTelemetryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/telemetry": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Run Telemetry",
        "operationId": "run_telemetry_runs__run_id__telemetry_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunTelemetryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/llm-calls": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "List Llm Calls",
        "operationId": "list_llm_calls_llm_calls_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Book Id"
            }
          },
          {
            "name": "chapter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "name": "scene_no",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scene No"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "name": "truncated",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Truncated"
            }
          },
          {
            "name": "errors_only",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Errors Only"
            }
          },
          {
            "name": "fallbacks_only",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fallbacks Only"
            }
          },
          {
            "name": "min_latency_ms",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Latency Ms"
            }
          },
          {
            "name": "min_input_tokens",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Input Tokens"
            }
          },
          {
            "name": "cache_miss_only",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cache Miss Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmCallListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/llm-calls/{call_id}": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Llm Call Detail",
        "operationId": "llm_call_detail_llm_calls__call_id__get",
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Call Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmCallOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}/telemetry/problems": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Book Telemetry Problems",
        "operationId": "book_telemetry_problems_books__book_id__telemetry_problems_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryProblemsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}/telemetry/compare": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Compare Runs",
        "operationId": "compare_runs_books__book_id__telemetry_compare_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          },
          {
            "name": "run_a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run A"
            }
          },
          {
            "name": "run_b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run B"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCompareOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookTelemetryOut": {
        "properties": {
          "totals": {
            "$ref": "#/components/schemas/TelemetryTotals",
            "default": {
              "calls": 0,
              "input_tokens": 0,
              "output_tokens": 0,
              "cache_creation_tokens": 0,
              "cache_read_tokens": 0,
              "cache_hit_ratio": 0.0,
              "cache_tokens_saved": 0,
              "truncations": 0,
              "errors": 0,
              "fallbacks": 0,
              "estimated_cost_usd": 0.0,
              "cache_savings_usd": 0.0
            }
          },
          "by_chapter": {
            "items": {
              "$ref": "#/components/schemas/ChapterRollupOut"
            },
            "type": "array",
            "title": "By Chapter",
            "default": []
          },
          "by_run": {
            "items": {
              "$ref": "#/components/schemas/RunRollupOut"
            },
            "type": "array",
            "title": "By Run",
            "default": []
          },
          "run_total": {
            "type": "integer",
            "title": "Run Total",
            "default": 0
          },
          "by_stage": {
            "items": {
              "$ref": "#/components/schemas/TelemetryGroupOut"
            },
            "type": "array",
            "title": "By Stage",
            "default": []
          },
          "by_model": {
            "items": {
              "$ref": "#/components/schemas/TelemetryGroupOut"
            },
            "type": "array",
            "title": "By Model",
            "default": []
          }
        },
        "type": "object",
        "title": "BookTelemetryOut",
        "description": "Global telemetry for a book: overall totals plus comparison rollups across chapters, stages,\nand models \u2014 the cross-chapter/scene view the global Telemetry tab renders."
      },
...
      "ChapterTelemetryOut": {
        "properties": {
          "chapter_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chapter Id"
          },
          "totals": {
            "$ref": "#/components/schemas/TelemetryTotals",
            "default": {
              "calls": 0,
              "input_tokens": 0,
              "output_tokens": 0,
              "cache_creation_tokens": 0,
              "cache_read_tokens": 0,
              "cache_hit_ratio": 0.0,
              "cache_tokens_saved": 0,
              "truncations": 0,
              "errors": 0,
              "fallbacks": 0,
              "estimated_cost_usd": 0.0,
              "cache_savings_usd": 0.0
            }
          },
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/SceneTelemetryOut"
            },
            "type": "array",
            "title": "Scenes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "chapter_id"
        ],
        "title": "ChapterTelemetryOut",
        "description": "Per-chapter derive telemetry: chapter totals + a per-scene breakdown."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LlmCallListOut": {
        "properties": {
          "calls": {
            "items": {
              "$ref": "#/components/schemas/LlmCallOut"
            },
            "type": "array",
            "title": "Calls",
            "default": []
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 50
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "LlmCallListOut"
      },
...
      "RunCompareOut": {
        "properties": {
          "run_a": {
            "$ref": "#/components/schemas/RunRollupOut"
          },
          "run_b": {
            "$ref": "#/components/schemas/RunRollupOut"
          },
          "stage_deltas": {
            "items": {
              "$ref": "#/components/schemas/StageDeltaOut"
            },
            "type": "array",
            "title": "Stage Deltas",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "run_a",
          "run_b"
        ],
        "title": "RunCompareOut"
      },
...
      "TelemetryProblemsOut": {
        "properties": {
          "problems": {
            "items": {
              "$ref": "#/components/schemas/TelemetryProblemOut"
            },
            "type": "array",
            "title": "Problems",
            "default": []
          },
          "healthy": {
            "type": "boolean",
            "title": "Healthy",
            "default": true
          }
        },
        "type": "object",
        "title": "TelemetryProblemsOut"
      },
...
    }
  }
}